+++ tests/test_functional_experiment.py
@@ -1,5 +1,7 @@
 import os
+import sys
 
+from pycomex.testing import ConfigIsolation
 from pycomex.testing import ExperimentIsolation
 from pycomex.functional.experiment import Experiment, run_experiment
 
@@ -18,20 +20,25 @@ def test_run_experiment_works():
     assert len(experiment.data) != 0
 
 
-def test_experiment_construction_basically_works():
-    """
-    A new Experiment object can be constructed without raising errors
-    """
-    value = 10
-    with ExperimentIsolation(glob_mod={'PARAMETER': value}) as iso:
-        experiment = Experiment(
-            base_path=iso.path,
-            namespace='experiment',
-            glob=iso.glob,
-        )
-        
-        assert 'PARAMETER' in experiment.parameters
-        assert experiment.parameters['PARAMETER'] == value
-        # Here we test if the alternative experiment parameter accessing api works
-        # Technically each parameter will automatically be exposed as a property
-        assert experiment.PARAMETER == value
+class TestExperiment:
+    
+    def test_construction_basically_works(self):
+        """
+        It should be possible to construct the experiment object without any errors and after 
+        constructing it should have the parameters that were passed to it.
+        """
+        parameters = {'PARAMETER': 10}
+        with ConfigIsolation() as config, ExperimentIsolation(sys.argv, glob_mod=parameters) as iso:
+            
+            config.load_plugins()
+            
+            experiment = Experiment(
+                base_path=iso.path,
+                namespace='experiment',
+                glob=iso.glob,
+            )
+            
+            assert isinstance(experiment, Experiment)
+            assert 'PARAMETER' in experiment.parameters
+
+